Godot 4 migration - automated conversion + basic C# fixes

+++ knytt/GDKnyttCamera.cs
@@ -29,7 +29,7 @@ public class GDKnyttCamera : Camera2D
         this.Target = target;
     }
 
-    public override void _PhysicsProcess(float delta)
+    public override void _PhysicsProcess(double delta)
     {
         if (delta <= 0) { GD.Print("Invalid delta: ", delta); return; }
